[onchain-discovery] move client out of lib
